requirements.txt
streamlit
pandas
sqlite3
gspread
oauth2client
psycopg2-binary
